+++ src/main/java/hello/servlet/servelt/ServletApplication.java
@@ -1,4 +1,4 @@
-package hello.servlet;
+package hello.servlet.servelt;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
